dxf: prepare codepage conversions, cquote
from cp 28 to 30 (czech to latin-1) we would need to quote much more
c>128 to \U+xxxx chars.
essentially all not representable chars in the current cp need to be
converted to \U+xxxx. So we really need the char tables for all our
codepages. which we could do with iconv or libc (btowc).

NOTE: we don't really need to do this. For ascii -> ascii conversion
keep the current codepages. Just UTF-8 conversions for json and
logging will be inaccurate with non 30 codepages, but the import will
keep the codepage and it will cancel out.
For ascii -> unicode same problem. Some btowc conversions need to done
for non-30 cp's
For unicode we simply emit to cp 30, which needs no conversions.